+++ rector.php
@@ -35,7 +35,7 @@ return RectorConfig::configure()
         SetList::INSTANCEOF,
         SetList::STRICT_BOOLEANS,
         // PHP-Unit
-        PHPUnitSetList::PHPUNIT_100,
+        PHPUnitSetList::PHPUNIT_110,
         PHPUnitSetList::PHPUNIT_CODE_QUALITY,
         PHPUnitSetList::ANNOTATIONS_TO_ATTRIBUTES,
     ]);